+++ CHANGELOG.md
@@ -10,6 +10,7 @@ we hit release version 1.0.0.
 
 ### Added
 - "Hz", "MHz", "GHz", "THz", and "invcm" as valid energy units, #725
+- added `read_gtensor` and `read_hyperfine_coupling` to `txtSileORCA`, #722
 - enabled `AtomsArgument` and `OrbitalsArgument` to accept `bool` for *all* or *none*
 - enabled `winSileWannier90.read_hamiltonian` to read the ``_tb.dat`` files
 - `atoms` argument to `DensityMatrix.spin_align` to align a subset of atoms